Restructured code to resemble a Python package, updated imports. Factored out entrypoint into main() so it's callable.
