+++ gthack/modules/extractInformation.py
