+++ gthack/modules/getSampleInfo.py
