+++ gthack/modules/write_gtc.py
