+++ gthack/modules/write_gtc_original.py
